Refactor enemy and bullet properties to use float types; update health handling and UI adjustments

+++ Assets/Scripts/BulletData.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 public class BulletData : ScriptableObject
 {
     public float speed = 10f;
-    public int damage = 10;
-    public float fireRate = 0.2f; // задержка между выстрелами
+    public float damage = 10;
+    public float fireRate = 0.2f; // пїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅ пїЅпїЅпїЅпїЅпїЅ пїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅпїЅ
     public float fireRange = 5f;
     [Range(0f, 100f)] public float accuracy = 0f;
     public int bulletsCount = 1;
